+++ inbm/Changelog.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Fixed
  - (NEXMANAGE-872) Fix provision-tc issue in TiberOS - cannot overwrite /etc/dispatcher.environment
 
+### Changed
+ - (NEXMANAGE-853) Remove ORAS and update method
+
 ## 4.2.6 - 2024-10-04
 ### Added
  - (NEXMANAGE-515) Update dispatcher SOTA related classes for supporting TiberOS